# lib/api-zod/src/generated/types/brand.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Wholesale mobile parts store API
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceModel } from './deviceModel';

export interface Brand {
  id: number;
  name: string;
  models: DeviceModel[];
}