# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Wholesale mobile parts store API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  id: number;
  productId: number;
  sku: string;
  name: string;
  quality: string;
  /** @nullable */
  imageUrl: string | null;
  quantity: number;
  unitPrice: number;
  lineTotal: number;
  stock: number;
}