# lib/api-zod/src/generated/types/catalogSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Wholesale mobile parts store API
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogSummary {
  productCount: number;
  brandCount: number;
  categoryCount: number;
  inStockCount: number;
}