# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Wholesale mobile parts store API
 * OpenAPI spec version: 0.1.0
 */
import type { NextTierProgress } from './nextTierProgress';
import type { OrderSummary } from './orderSummary';
import type { PriceTier } from './priceTier';

export interface DashboardSummary {
  totalOrders: number;
  annualSpend: number;
  totalSavings: number;
  tier: PriceTier;
  nextTier: NextTierProgress | null;
  recentOrders: OrderSummary[];
}